split notebook classes so they only implement one interface each
